Change idref_t to be a pointer to the start of the string rather than     make a copy.  It was a copy since it was 'easier', but all those copies     got real expensive when you're doing hundreds of millions of them.
Add a switch across the first letter before running the gauntlet of the
    MATCH()'s, much more efficient until the whole thing can be replaced
    by a real grammar.

Take advantage of knowing the ident length in every place I saw.  We have
    had that for quite a while, just weren't taking great advantage of it.

Change the holder_t struct from a malloc to a tmalloc, much better use of
    memory.

Fix an actual bug in what object we mark as done in the hash.
    NOTE: This should be back ported to 1.1.x.. though I think its harmless
    as objnum == -1 in both of those cases.  I haven't really assessed the
    impact of it, but obviously its not a HUGE problem or people would have
    reported dataloss.

Split up dump_object into 3 functions.

Drop some code that's been #if 0'ed for ages.