Remove system(rm -f runfile).
If the unlink fails, we don't need to fall back to doing a `system()`
to invoke `rm -f`. There's no clear history on why this was added,
so I'm chalking it up to "something was probably odd on some Unix
variant back in the mid-1990s" rather than anything that should
be necessary today.