Begin switching to newer test infrastructure.
This is from Jeremy Parker (I think) who had it in his repo at
https://github.com/nrhtr/genesis

I've got this currently hardcoded to assume that `coldcc` is in
a `build` directory which will have to be improved.

Moving forward, we will want to move the various tests from the
legacy arrangement into the new system and then improve the
coverage.