binarydb: Switch to non-buffered I/O.
Since we know exactly what we want to read (and write), and aren't
usually doing sequential I/O, there's not much point in using the
buffered libc I/O code.

Compiling a very large DB has little noticeable change from this,
but it should be a more useful thing at runtime on a large DB
under load as it may save I/O by only doing what's needed.

Previously, this did call `fflush`, which would flush the buffered
I/O out to the OS. This isn't the same as calling `fsync` or friends,
so we don't do that here.

Fixes #19.