Cast objnums to long when printing.
When running with `USE_BIG_NUMBERS`, `cObjnum` is `long`, so we
should print it with `%ld` instead of `%d`.

To keep things easy, we'll just always print it as `%ld` and cast
it when passing to `printf`. This lets it work in builds with
and without the `USE_BIG_NUMBERS` flag set without generating
warnings.