Add FLOAT_ZERO, FLOAT_ONE opcodes.
This is a (minor) optimization for size in method bytecode
by adding specialized ops for representing `0.0` and `1.0`.

This is similar to the existing opcodes for `0` and `1` for
integers.

For floats, since we store the float as the bits of the value
as an integer, these are values that end up not compressing
well with `write_long` since all of the 32 (or 64) bits are
required. Since `0.0` and `1.0` are relatively common, having
something for them can save in the size of the compressed
opcodes, even if in a relatively minor way.

Fixes #18.